Switched to function based view for adding new chains
I switched to a FBV because the model form set has
a dynamic number of forms determined by the query
param 'num_chains'. Since the model formset factory
is used both in the get and in the post calls, I
thought it was better to just create it once and use
it for both request methods.