Add API for retrieving game info
I'm planning to retrive game info in inpector app via AJAX
instead of templating it's json representation. I think
that this way should be easier to integrate with
Backbone-powered inspector app.

Also, I've changed the URL schema for inpector API a bit:
now it's not bound to the game id.