Add workaround to avoid views unnecessary rerendering
In some case we do not want for view to redraw itself in response
to model change because that change was initiated by view itself
and it's UI already in proper state, but still want to trigger
model `change` event in order to redraw other views for that model.

To achieve this, I've added boolean flags to such views that
indicates whether the render is self-initiated or not.